requirements.txt
Django==4.1.5
djangorestframework==3.14.0
whitenoise==3.3.1
pandas==1.1.4